order is created  and swagger is implemented in railway and render

+++ services/PrdoctService.js
@@ -1,5 +1,5 @@
 const Product = require("../models/ProductSchema");
-
+const User = require("../models/userSchema");
 exports.createProduct = async (
   Title,
   Description,
@@ -9,18 +9,23 @@ exports.createProduct = async (
   owner
 ) => {
   try {
-    console.log("in service product");
-    const NewProduct = new Product(
+    console.log("in service product", owner);
+    const userRole = await User.findOne({ _id: owner });
+    console.log(userRole.role);
+    if (userRole.role == "Consumer")
+      return { message: "user with consumer role cannot create product" };
+    const NewProduct = new Product({
       Title,
       Description,
       Price,
       Availability,
       Category,
-      owner
-    );
+      owner,
+    });
     await NewProduct.save();
     return NewProduct;
   } catch (error) {
+    console.log(error);
     throw { message: error.message };
   }
 };
@@ -40,10 +45,18 @@ exports.updateProduct = async (
   Description,
   Price,
   Availability,
-  Category
+  Category,
+  userId
 ) => {
   try {
     console.log("in service");
+    const user = await User.findOne({ _id: userId });
+    console.log(user.role);
+    if (user.role == "Consumer")
+      return {
+        message:
+          "user exists , but Consumer is not authorized to do this action",
+      };
     const product = await Product.findOneAndUpdate(
       { _id: pid },
       { Title, Description, Price, Availability, Category },
@@ -55,9 +68,16 @@ exports.updateProduct = async (
   }
 };
 
-exports.deleteProduct = async (id) => {
+exports.deleteProduct = async (userid, id) => {
   try {
     console.log("in service");
+    const user = await User.findOne({ _id: userid });
+    console.log(user.role);
+    if (user.role == "Consumer")
+      return {
+        message:
+          "user exists , but Consumer is not authorized to do this action",
+      };
     const product = await Product.findOneAndDelete({ _id: id });
     return product;
   } catch (error) {
